feat: v1.0.2 — B1 均衡器EQ + B5 睡眠定时器 + D3 歌词精度升级(Float→Long)
B1 均衡器 (Equalizer):
- 新增 EqualizerManager.kt：封装 Android Equalizer API，6 套内置预设（默认/流行/古典/电子/人声增强/低音增强），try-catch 防止设备不兼容崩溃
- 新增 EqualizerScreen.kt：均衡器页面，预设选择 + 5 频段垂直 Slider，配色跟随 HSL 主题
- PlaybackService 在 ExoPlayer 初始化后立即绑定 EqualizerManager，onDestroy 时自动释放
- MainViewModel 新增 eqPreset/eqBandLevels 状态、setEqPreset/setEqBandLevel 函数（持久化到 DataStore）
- SettingsManager 新增 KEY_EQ_PRESET/KEY_EQ_BAND_LEVELS 持久化键
- SettingsScreen '播放控制' 分区新增 EQ 入口，显示当前预设名
- MainActivity 注册 EqualizerScreen 路由（右滑进入，系统返回键/手势可退出）

B5 睡眠定时器 (Sleep Timer):
- 新增 SleepTimerDialog.kt：Material3 弹窗，支持 15/30/45/60 分钟 + 当前曲结束后停止 + 取消定时
- MainViewModel 新增 sleepTimerRemaining/sleepAfterCurrentTrack 状态，setSleepTimer/setSleepAfterCurrentTrack/cancelSleepTimer 函数
- SettingsScreen 在 '播放控制' 分区展示睡眠定时器入口，实时显示剩余倒计时
- onMediaItemTransition 监听支持「当前曲结束」模式自动暂停

D3 歌词精度升级 (Lyric Precision):
- LyricLine.time: Float(秒) → Long(毫秒)，彻底消除浮点转换误差
- LyricParser 直接计算毫秒：(minutes * 60 + seconds) * 1000L + milliseconds
- MainViewModel.updateLyricIndex() 移除 /1000f 转换，直接 Long 比较
- LyricsPanel 点击跳转由 (line.time * 1000).toLong() 改为 line.time（直接传毫秒）

Version: 1.0.1 → 1.0.2 (versionCode 101 → 102)

+++ android-native/app/src/main/java/com/openmusic/app/data/LyricLine.kt
@@ -1,6 +1,6 @@
 package com.openmusic.app.data
 
 data class LyricLine(
-    val time: Float, // Time offset in seconds
+    val time: Long, // Time offset in milliseconds (Long for precision, eliminates Float rounding errors)
     val text: String
 )
